Move PresetOverviewScreen to separate module
